211 - Adding column types in the WFOBasicTable component and moving the depending components in the same folder
