+++ packages/orchestrator-ui-components/src/components/WFOTable/WFOBasicTable/index.ts
@@ -1,1 +1,3 @@
 export * from './WFOBasicTable';
+export * from './WFOSortDirectionIcon';
+export * from './WFOTableHeaderCell';
